+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### ❤️ New Stuff
 
 - Added packet rate logging to deauth attacks with 5s intervals - @jaylikesbunda
-- Added 'rgbmode' command to control the RGB LED strip directly - @jaylikesbunda
+- Added 'rgbmode' command to control the RGB LEDs directly with support for color and mode args- @jaylikesbunda
+- Added new 'strobe' effect for RGB LEDs - @jaylikesbunda
 
 ### 🤏 Tweaks and Improvements